Logica de negocio del modulo usuarios

+++ src/main/java/com/example/DDP/Usuarios/Repository/ProveedorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProveedorRepository extends CrudRepository<Proveedor, Long> {
+public interface ProveedorRepository extends CrudRepository<Proveedor, Integer> {
 
 }
